changes
Created Move class, added more code to AI.java. changed the visibility
of the validmove method to public. added a setCurrentChecker method to
board.java to be used in the AI class to help populate list of possible
moves. Code for generating possible moves for regular red and black
pieces should be working